Move definitions of EmailAddress, EmailType into Events module
